Use int rather than allowing decimal values. Add a nicer looking selector field

+++ packages/validators/src/tasting-notes.validator.ts
@@ -2,13 +2,13 @@ import { z } from "zod";
 
 export const createTastingNoteSchema = z.object({
   vintageId: z.number().int().positive(),
-  score: z.number().min(0).max(10),
+  score: z.number().int().min(0).max(10),
   notes: z.string().trim(),
 });
 
 export const updateTastingNoteSchema = z
   .object({
-    score: z.number().min(0).max(10).optional(),
+    score: z.number().int().min(0).max(10).optional(),
     notes: z.string().trim().optional(),
   })
   .refine((data) => Object.keys(data).length > 0, {
